separate phone component, copy via props
